# feat: optimize reading a workspace's files