--- a/tooling/nargo/src/lib.rs
+++ b/tooling/nargo/src/lib.rs
@@ -23,6 +23,7 @@ use noirc_frontend::{
 };
 use package::{Dependency, Package};
 use rayon::prelude::*;
+use walkdir::WalkDir;
 
 pub use self::errors::NargoError;
 
@@ -79,16 +80,28 @@ fn insert_all_files_for_package_into_file_manager(
         .parent()
         .unwrap_or_else(|| panic!("The entry path is expected to be a single file within a directory and so should have a parent {:?}", package.entry_path));
 
-    // Get all files in the package and add them to the file manager
-    let paths = get_all_noir_source_in_dir(entry_path_parent)
-        .expect("could not get all paths in the package");
-    for path in paths {
+    for entry in WalkDir::new(entry_path_parent) {
+        let Ok(entry) = entry else {
+            continue;
+        };
+
+        if !entry.file_type().is_file() {
+            continue;
+        }
+
+        if !entry.path().extension().map_or(false, |ext| ext == FILE_EXTENSION) {
+            continue;
+        };
+
+        let path = entry.into_path();
+
         let source = if let Some(src) = overrides.get(path.as_path()) {
             src.to_string()
         } else {
             std::fs::read_to_string(path.as_path())
                 .unwrap_or_else(|_| panic!("could not read file {:?} into string", path))
         };
+
         file_manager.add_file_with_source(path.as_path(), source);
     }
 
@@ -142,85 +155,4 @@ pub fn prepare_package<'file_manager, 'parsed_files>(
     prepare_dependencies(&mut context, crate_id, &package.dependencies);
 
     (context, crate_id)
-}
-
-// Get all Noir source files in the directory and subdirectories.
-//
-// Panics: If the path is not a path to a directory.
-fn get_all_noir_source_in_dir(dir: &std::path::Path) -> std::io::Result<Vec<std::path::PathBuf>> {
-    get_all_paths_in_dir(dir, |path| {
-        path.extension().map_or(false, |extension| extension == FILE_EXTENSION)
-    })
-}
-
-// Get all paths in the directory and subdirectories.
-//
-// Panics: If the path is not a path to a directory.
-//
-// TODO: Along with prepare_package, this function is an abstraction leak
-// TODO: given that this crate should not know about the file manager.
-// TODO: We can clean this up in a future refactor
-fn get_all_paths_in_dir(
-    dir: &std::path::Path,
-    predicate: fn(&std::path::Path) -> bool,
-) -> std::io::Result<Vec<std::path::PathBuf>> {
-    assert!(dir.is_dir(), "directory {dir:?} is not a path to a directory");
-
-    let mut paths = Vec::new();
-
-    if dir.is_dir() {
-        for entry in std::fs::read_dir(dir)? {
-            let entry = entry?;
-            let path = entry.path();
-            if path.is_dir() {
-                let mut sub_paths = get_all_paths_in_dir(&path, predicate)?;
-                paths.append(&mut sub_paths);
-            } else if predicate(&path) {
-                paths.push(path);
-            }
-        }
-    }
-
-    Ok(paths)
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::get_all_paths_in_dir;
-    use std::{
-        fs::{self, File},
-        path::Path,
-    };
-    use tempfile::tempdir;
-
-    fn create_test_dir_structure(temp_dir: &Path) -> std::io::Result<()> {
-        fs::create_dir(temp_dir.join("sub_dir1"))?;
-        File::create(temp_dir.join("sub_dir1/file1.txt"))?;
-        fs::create_dir(temp_dir.join("sub_dir2"))?;
-        File::create(temp_dir.join("sub_dir2/file2.txt"))?;
-        File::create(temp_dir.join("file3.txt"))?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_all_paths_in_dir() {
-        let temp_dir = tempdir().expect("could not create a temporary directory");
-        create_test_dir_structure(temp_dir.path())
-            .expect("could not create test directory structure");
-
-        let paths = get_all_paths_in_dir(temp_dir.path(), |_| true)
-            .expect("could not get all paths in the test directory");
-
-        // This should be the paths to all of the files in the directory and the subdirectory
-        let expected_paths = vec![
-            temp_dir.path().join("file3.txt"),
-            temp_dir.path().join("sub_dir1/file1.txt"),
-            temp_dir.path().join("sub_dir2/file2.txt"),
-        ];
-
-        assert_eq!(paths.len(), expected_paths.len());
-        for path in expected_paths {
-            assert!(paths.contains(&path));
-        }
-    }
 }